Add configuration file
In order to keep secret information out of source control, adding a
configuration file. Configuration file is in the .gitignore file so that
it does not accidentally get checked in. A default configuration file
devoid of secret information is in the config.tpl.js file. A check to
ensure that configurations we know need to be changed has been added.