Move books route into file
Moved the books route out of the index.js file and into its own file for
importing. This also required that the aws-client and mong-cache-control
become 'unclassed' and separated into separate export functions to be
shared across the modules.

Moving books into its own files allows for easy addition of more routes
in the future.